Removed AudioDataSource (for now). Changed all int64_t and long long types to SInt64. Started adding Buffered data source

+++ StreamingKit/StreamingKit/STKAutoRecoveringHTTPDataSource.m
@@ -246,7 +246,7 @@ static void PopulateOptionsWithDefault(STKAutoRecoveringHTTPDataSourceOptions* o
     return NO;
 }
 
--(void) seekToOffset:(long long)offset
+-(void) seekToOffset:(int64_t)offset
 {
 	ticksWhenLastDataReceived = GetTickCount();
